Mario Attempt #1
I think I fixed the choosing colour issue with over saturation but I had
to revert to the original scipt and see what was happening.  The main
issue was colour needs to be normalized from 0-1 in webGL but the colour
picker was choosing from 0-255.  Easy fix is normalize before using.